requirements.txt
load-dotenv==0.1.0
pillow==11.0.0
python-dotenv==1.0.1
tqdm==4.67.0
